+++ NEWS.md
@@ -10,6 +10,10 @@ None
 -  `ts_acceleration_vec()`
 -  `ts_acceleration_augment()`
 -  `step_ts_accleration()`
+4. Fix #155 - Add functions:
+-  `ts_velocity_vec()`
+-  `ts_velocity_augment()`
+-  `step_ts_velocity()`
 
 ## Minor Fixes and Improvements
 1. Fix #159 - Add parameter `.date_col` to `ts_sma_plot()` so that if a tibble is passed